+++ src/com/scs/moonbaseassault/client/modules/IntroJonlan.java
@@ -59,7 +59,7 @@ public class IntroJonlan extends AbstractModule {//implements ActionListener {
 		client.getCamera().setLocation(new Vector3f(0, 0.55f, 0));
 		client.getCamera().lookAt(new Vector3f(10, 0.55f, 10), Vector3f.UNIT_Y);
 
-		jonlanModel = new SoldierModel(client.getAssetManager(), 1, false, false);
+		jonlanModel = new SoldierModel(client.getAssetManager(), (byte)1, false, false);
 		jonlanModel.createAndGetModel();
 		jonlanModel.setAnim(AbstractAvatar.ANIM_RUNNING);
 		jonlanModel.getModel().setLocalTranslation(10, 0, 5);
