+++ src/com/scs/moonbaseassault/entities/MA_AISuperSoldier.java
@@ -9,7 +9,7 @@ import com.scs.stevetech1.shared.IEntityController;
 
 public class MA_AISuperSoldier extends AbstractAISoldier {
 	
-	public MA_AISuperSoldier(IEntityController _game, int id, float x, float y, float z, int _side, boolean friend, int csInitialAnimCode, String name) {
+	public MA_AISuperSoldier(IEntityController _game, int id, float x, float y, float z, byte _side, boolean friend, int csInitialAnimCode, String name) {
 		super(_game, id, MoonbaseAssaultClientEntityCreator.AI_SUPER_SOLDIER, x, y, z, _side, 
 				new SoldierModel(_game.getAssetManager(), _side, false, friend), csInitialAnimCode, name);
